Fixes:
- fix type when filtering product by price met target.
- fix the lowest price was overridden if price wasn't crawled.
- fix store tabs not working
- fix some european amazon using "." as thousands separator.
- keep last 3 days of logs only.
- added volume to .env to fix 2FA not working after restarting container
- added volume for database
-

New Stores:
- best buy

New Features:

- add other users to the system and manage their tokens.
- add telegram notification.
- add percentage drop.
- robot detection for amazon ( if the system was detected as robot, it will stop executing and try the next product, the same product that was blocked will be crawled again in the next batch)

Took 6 hours 50 minutes

+++ docker/entrypoint.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-cp .env.example .env
-
+if [ ! -f ".env" ]; then
+    cp .env.example .env
+    php artisan key:generate --force
+fi
 
 php artisan storage:link
 php artisan config:clear
@@ -12,8 +14,7 @@ php artisan key:generate --force
 
 printenv > /etc/environment
 
-php artisan migrate --seed --force
-
+php artisan migrate  --force --seed
 
 php artisan discount:fill-supervisor-workers
 
@@ -24,7 +25,6 @@ php artisan make:filament-user --name=$DEFAULT_USER --email=$DEFAULT_EMAIL --pas
 php artisan octane:install --server=frankenphp
 
 Xvfb :99 -screen 0 2000x2000x24 & export DISPLAY=:99
-
 
 supervisord -c  /etc/supervisor/conf.d/supervisord.conf
